# Improve the performance of truncating existing hdf5 files